Avoid implicit double dtype conversion

+++ scipy/stats/stats.py
@@ -5194,7 +5194,7 @@ def kruskal(*args, **kwargs):
     for i in range(num_groups):
         ssbn += _square_of_sums(ranked[j[i]:j[i+1]]) / float(n[i])
 
-    totaln = np.sum(n, dtype='uint64')
+    totaln = np.sum(n, dtype=float)
     h = 12.0 / (totaln * (totaln + 1)) * ssbn - 3 * (totaln + 1)
     df = num_groups - 1
     h /= ties
